refactor(Disclaimer, Icon): update SchemeColor import path for consistency and re-export for backward compatibility
