+++ src/components/Icon/types.ts
@@ -1,91 +1,7 @@
-export type SchemeColor = 
-  | 'background-solo'
-  | 'background-sky'
-  | 'background-grass'
-  | 'background-carpet'
-  | 'background-snow'
-  | 'background-lightest'
-  | 'background-white'
-  | 'neutral-black'
-  | 'neutral-darkest'
-  | 'neutral-dark'
-  | 'neutral-cloudy'
-  | 'neutral-cleanest'
-  | 'neutral-clean'
-  | 'neutral-soft'
-  | 'neutral-lightest'
-  | 'neutral-light'
-  | 'neutral-white'
-  | 'neutral-snow'
-  | 'feedback-red'
-  | 'feedback-green'
-  | 'feedback-yellow'
-  | 'feedback-blue'
-  | 'feedback-grey'
-  | 'aux-blue'
-  | 'aux-purple'
-  | 'aux-orange'
-  | 'aux-lemon'
-  | 'aux-pink'
-  | 'aux-baby-blue'
-  | 'aux-baby-yellow'
-  | 'aux-baby-red'
-  | 'aux-baby-green'
-  | 'aux-baby-pink'
-  | 'aux-green-100'
-  | 'aux-green-300'
-  | 'aux-green-500'
-  | 'aux-green-700'
-  | 'aux-green-900'
-  | 'aux-blue-100'
-  | 'aux-blue-300'
-  | 'aux-blue-500'
-  | 'aux-blue-700'
-  | 'aux-blue-900'
-  | 'aux-purple-100'
-  | 'aux-purple-300'
-  | 'aux-purple-500'
-  | 'aux-purple-700'
-  | 'aux-purple-900'
-  | 'aux-red-100'
-  | 'aux-red-300'
-  | 'aux-red-500'
-  | 'aux-red-700'
-  | 'aux-red-900'
-  | 'aux-orange-100'
-  | 'aux-orange-300'
-  | 'aux-orange-500'
-  | 'aux-orange-700'
-  | 'aux-orange-900'
-  | 'aux-yellow-100'
-  | 'aux-yellow-300'
-  | 'aux-yellow-500'
-  | 'aux-yellow-700'
-  | 'aux-yellow-900'
-  | 'floweditor-blue'
-  | 'floweditor-purple'
-  | 'floweditor-wine'
-  | 'floweditor-orange'
-  | 'floweditor-pink'
-  | 'floweditor-turquoise'
-  | 'floweditor-green'
-  | 'weni-50'
-  | 'weni-100'
-  | 'weni-200'
-  | 'weni-300'
-  | 'weni-400'
-  | 'weni-500'
-  | 'weni-600'
-  | 'weni-700'
-  | 'weni-800'
-  | 'weni-900'
-  | 'weni-950'
-  | 'brand-weni'
-  | 'brand-weni-dark'
-  | 'brand-weni-soft'
-  | 'brand-sec-dark'
-  | 'brand-sec-soft'
-  | 'brand-sec';
+import type { SchemeColor } from "@/types/scheme-colors";
+
+// Re-export SchemeColor for backward compatibility
+export type { SchemeColor };
 
 export type IconSize = 
   | 'nano'
